backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart